+++ gradle.properties
@@ -2,14 +2,14 @@
 org.gradle.jvmargs=-Xmx1G
 
 # Fabric Properties https://fabricmc.net/versions.html
-minecraft_version=1.19.4
-yarn_mappings=1.19.4+build.1
-loader_version=0.14.19
+minecraft_version=1.20
+yarn_mappings=1.20+build.1
+loader_version=0.14.21
 
 # Dependencies
-fabric_version=0.76.0+1.19.4
-modmenu_version=6.1.0-rc.3
-cloth_config_2_version=10.0.96
+fabric_version=0.83.0+1.20
+modmenu_version=7.0.1
+cloth_config_2_version=11.0.99
 
 # Mod Properties
 mod_version=1.4.3
