+++ src/main/java/dev/bernasss12/bebooks/client/gui/tooltip/IconTooltipComponent.java
@@ -1,8 +1,8 @@
 package dev.bernasss12.bebooks.client.gui.tooltip;
 
 import net.minecraft.client.font.TextRenderer;
+import net.minecraft.client.gui.DrawContext;
 import net.minecraft.client.gui.tooltip.TooltipComponent;
-import net.minecraft.client.render.item.ItemRenderer;
 import net.minecraft.client.util.math.MatrixStack;
 import net.minecraft.item.ItemStack;
 
@@ -21,15 +21,16 @@ public record IconTooltipComponent(List<ItemStack> icons) implements TooltipComp
     }
 
     @Override
-    public void drawItems(TextRenderer textRenderer, int x, int y, MatrixStack matrices, ItemRenderer itemRenderer) {
+    public void drawItems(TextRenderer textRenderer, int x, int y, DrawContext context) {
         float scale = 0.5f;
         int scaledX = (int) (x / scale);
         int scaledY = (int) (y / scale);
         int scaledOffset =  (int) (8 / scale);
+        MatrixStack matrices = context.getMatrices();
         matrices.push();
         matrices.scale(0.5f, 0.5f, 1.0f);
-        for(int i = 0; i < icons.size(); i++){
-            itemRenderer.renderInGuiWithOverrides(matrices, icons.get(i), scaledX + scaledOffset * i, scaledY, -1);
+        for (int i = 0; i < icons.size(); i++) {
+            context.drawItem(icons.get(i), scaledX + scaledOffset * i, scaledY);
         }
         matrices.pop();
     }
