+++ src/main/java/dev/bernasss12/bebooks/mixin/HandledScreenMixin.java
@@ -1,0 +1,51 @@
+package dev.bernasss12.bebooks.mixin;
+
+import dev.bernasss12.bebooks.BetterEnchantedBooks;
+import dev.bernasss12.bebooks.client.gui.ModConfig;
+import dev.bernasss12.bebooks.client.gui.tooltip.IconTooltipComponent;
+import dev.bernasss12.bebooks.util.text.IconTooltipDataText;
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+import net.minecraft.client.font.TextRenderer;
+import net.minecraft.client.gui.DrawContext;
+import net.minecraft.client.gui.screen.Screen;
+import net.minecraft.client.gui.screen.ingame.HandledScreen;
+import net.minecraft.client.gui.tooltip.OrderedTextTooltipComponent;
+import net.minecraft.client.gui.tooltip.TooltipComponent;
+import net.minecraft.client.util.math.MatrixStack;
+import net.minecraft.item.ItemStack;
+import net.minecraft.item.Items;
+import net.minecraft.text.OrderedText;
+import net.minecraft.text.Text;
+import org.apache.logging.log4j.LogManager;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.ModifyArg;
+import org.spongepowered.asm.mixin.injection.ModifyVariable;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Mixin(HandledScreen.class)
+@Environment(EnvType.CLIENT)
+public abstract class HandledScreenMixin {
+    @ModifyArg(
+        method = "drawMouseoverTooltip(Lnet/minecraft/client/gui/DrawContext;II)V",
+        at = @At(
+            value = "INVOKE",
+            target = "Lnet/minecraft/client/gui/screen/ingame/HandledScreen;getTooltipFromItem(Lnet/minecraft/item/ItemStack;)Ljava/util/List;"
+        ),
+        index = 0
+    )
+    private ItemStack setEnchantedItemStack(ItemStack stack) {
+        BetterEnchantedBooks.enchantedItemStack.set(stack);
+        return stack;
+    }
+
+    @Inject(method = "drawMouseoverTooltip(Lnet/minecraft/client/gui/DrawContext;II)V", at = @At(value = "TAIL"))
+    private void forgetEnchantedItemStack(DrawContext context, int x, int y, CallbackInfo ci) {
+        BetterEnchantedBooks.enchantedItemStack.set(ItemStack.EMPTY);
+    }
+}
